+++ client/src/app/models/tag-field.ts
@@ -530,6 +530,9 @@ export class VideoRecordField extends TagField {
   /**
    * The video record field requires a dataset to record to. This will provide
    * the user creating the form with a list of datasets to choose from.
+   *
+   * TODO: Need to validate that the minimumRequired is not greater than the
+   * maximumOptional
    */
   async getFieldSpecificProperties(): Promise<{
     [property: string]: JsonSchema;
@@ -547,6 +550,16 @@ export class VideoRecordField extends TagField {
         oneOf: options,
         description: 'The dataset to save the videos into',
       },
+      minimumRequired: {
+        type: 'number',
+        description:
+          'The minimum number of videos the user needs to record, (defaults to 1)',
+      },
+      maximumOptional: {
+        type: 'number',
+        description:
+          'The maximum number of videos the user can record (including required, defaults to 1)',
+      },
     };
   }
 
@@ -556,11 +569,38 @@ export class VideoRecordField extends TagField {
         type: 'Control',
         scope: '#/properties/dataset',
       },
+      {
+        type: 'Control',
+        scope: '#/properties/minimumRequired',
+        options: {
+          showUnfocusedDescription: true,
+        },
+      },
+      {
+        type: 'Control',
+        scope: '#/properties/maximumOptional',
+        options: {
+          showUnfocusedDescription: true,
+        },
+      },
     ];
   }
 
   protected getRequiredFieldProperties(): string[] {
     return ['dataset'];
+  }
+
+  asDataProperty(): JsonSchema {
+    return {
+      [this.getFieldName()]: {
+        type: 'array',
+        description: this.getDescription(),
+        items: {
+          type: 'string',
+        },
+        minItems: this.data.minimumRequired || 1,
+      },
+    };
   }
 
   asUIProperty(): any[] {
@@ -571,6 +611,8 @@ export class VideoRecordField extends TagField {
         options: {
           customType: 'video',
           dataset: this.data.dataset,
+          minimumRequired: this.data.minimumRequired || 1,
+          maximumOptional: this.data.maximumOptional || 1,
           showUnfocusedDescription: true,
         },
       },